FIT-486 fix major and minor code smells.
