+++ src/main/java/gov/ca/cwds/idm/service/cognito/AuditProperties.java
@@ -5,33 +5,6 @@ import org.springframework.context.annotation.Profile;
 
 @Profile("idm")
 @ConfigurationProperties(prefix = "audit")
-public class AuditProperties {
+public class AuditProperties extends CommonDoraProperties{
 
-  private String doraUrl;
-  private String index;
-  private String type;
-
-  public String getDoraUrl() {
-    return doraUrl;
-  }
-
-  public void setDoraUrl(String doraUrl) {
-    this.doraUrl = doraUrl;
-  }
-
-  public String getIndex() {
-    return index;
-  }
-
-  public void setIndex(String index) {
-    this.index = index;
-  }
-
-  public String getType() {
-    return type;
-  }
-
-  public void setType(String type) {
-    this.type = type;
-  }
 }
