+++ src/main/java/gov/ca/cwds/idm/service/cognito/CommonDoraProperties.java
@@ -1,0 +1,32 @@
+package gov.ca.cwds.idm.service.cognito;
+
+public abstract class CommonDoraProperties {
+
+  private String doraUrl;
+  private String index;
+  private String type;
+
+  public String getDoraUrl() {
+    return doraUrl;
+  }
+
+  public void setDoraUrl(String doraUrl) {
+    this.doraUrl = doraUrl;
+  }
+
+  public String getIndex() {
+    return index;
+  }
+
+  public void setIndex(String index) {
+    this.index = index;
+  }
+
+  public String getType() {
+    return type;
+  }
+
+  public void setType(String type) {
+    this.type = type;
+  }
+}
