+++ src/main/java/gov/ca/cwds/idm/service/notification/NotificationTypes.java
@@ -2,8 +2,10 @@ package gov.ca.cwds.idm.service.notification;
 
 public class NotificationTypes {
 
+  public static final String USER_LOCKED = "locked";
+
   private NotificationTypes() {
     throw new IllegalStateException("Notification Types");
   }
-  public static final String USER_LOCKED = "locked";
+
 }
