+++ src/main/java/gov/ca/cwds/rest/api/domain/auth/AuthPrivilege.java
@@ -22,7 +22,7 @@ public enum AuthPrivilege {
 
   private final String code;
   private final String description;
-
+  private static final Map<String, AuthPrivilege> mapByAuthPrivilegeCode = new HashMap<>();
 
   private AuthPrivilege(String code, String description) {
     this.code = code;
@@ -42,8 +42,6 @@ public enum AuthPrivilege {
   public String getDescription() {
     return description;
   }
-
-  private static final Map<String, AuthPrivilege> mapByAuthPrivilegeCode = new HashMap<>();
 
   public static AuthPrivilege findByAuthPrivilegeCode(String code) {
     return mapByAuthPrivilegeCode.get(code);
