+++ src/main/java/gov/ca/cwds/service/IdentityMappingService.java
@@ -3,7 +3,6 @@ package gov.ca.cwds.service;
 import javax.script.ScriptException;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import gov.ca.cwds.PerryProperties;
 import gov.ca.cwds.UniversalUserToken;
