+++ src/main/java/gov/ca/cwds/service/mfa/CognitoResponseService.java
@@ -5,7 +5,6 @@ import gov.ca.cwds.service.mfa.model.AuthParameters;
 import gov.ca.cwds.service.mfa.model.CognitoResponse;
 import gov.ca.cwds.service.mfa.model.RefreshRequest;
 import gov.ca.cwds.service.mfa.model.RefreshResponse;
-import io.swagger.models.RefResponse;
 import java.io.IOException;
 import java.util.Date;
 import java.util.HashMap;
@@ -17,7 +16,6 @@ import org.springframework.security.oauth2.client.DefaultOAuth2ClientContext;
 import org.springframework.security.oauth2.client.OAuth2ClientContext;
 import org.springframework.security.oauth2.common.DefaultOAuth2AccessToken;
 import org.springframework.security.oauth2.common.DefaultOAuth2RefreshToken;
-import org.springframework.security.oauth2.common.OAuth2AccessToken;
 import org.springframework.security.oauth2.common.OAuth2RefreshToken;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.stereotype.Service;
